fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@15242 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIVehicle.cpp
@@ -1376,7 +1376,7 @@ GUIVehicle::drawRouteHelper(const MSRoute& r, SUMOReal exaggeration) const {
 
 MSLane*
 GUIVehicle::getPreviousLane(MSLane* current, int& furtherIndex) const {
-    if (furtherIndex < myFurtherLanes.size()) {
+    if (furtherIndex < (int)myFurtherLanes.size()) {
         return myFurtherLanes[furtherIndex++];
     } else {
         return current;
